backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-storage==2.10.0
celery==5.3.4
redis==5.0.1
pymupdf==1.23.8
python-dotenv==1.0.0
PyJWT==2.8.1
pytesseract==0.3.10
Pillow==10.1.0
openai==0.27.8
